[FEATURE] Graceful handling of removed node types
Changes the error messages in the Backend, if a Node Type or a Document Type cannot be found.
Instead of showing two error messages with no hint on what's actually wrong, we now show only one
error message that shows the name of the missing node type.

Change-Id: I146813476a65388d9bc37b414013ce6c525667ae
Relates: NEOS-649
Release: master